requirements.txt
keras==2.11.0
tensorflow==2.11.0
transformers==4.30.2
sentence-transformers==2.2.2
huggingface_hub==0.14.1
faiss-cpu
streamlit
python-dotenv
google-generativeai
PyMuPDF
